+++ server/src/internalClusterTest/java/org/opensearch/remotestore/RemoteStoreIT.java
@@ -10,8 +10,6 @@ package org.opensearch.remotestore;
 
 import org.opensearch.OpenSearchException;
 import org.opensearch.action.admin.cluster.health.ClusterHealthResponse;
-import org.opensearch.action.admin.cluster.node.stats.NodeStats;
-import org.opensearch.action.admin.cluster.node.stats.NodesStatsResponse;
 import org.opensearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
 import org.opensearch.action.admin.cluster.snapshots.restore.RestoreSnapshotResponse;
 import org.opensearch.action.admin.indices.delete.DeleteIndexRequest;
@@ -40,7 +38,6 @@ import org.opensearch.indices.RemoteStoreSettings;
 import org.opensearch.indices.recovery.PeerRecoveryTargetService;
 import org.opensearch.indices.recovery.RecoverySettings;
 import org.opensearch.indices.recovery.RecoveryState;
-import org.opensearch.node.remotestore.RemoteStorePinnedTimestampService;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.test.InternalTestCluster;
 import org.opensearch.test.OpenSearchIntegTestCase;
@@ -62,7 +59,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.opensearch.action.admin.cluster.node.stats.NodesStatsRequest.Metric.REMOTE_STORE_NODE_STATS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_REPLICAS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_SHARDS;
 import static org.opensearch.index.remote.RemoteStoreEnums.DataCategory.SEGMENTS;
@@ -1015,103 +1011,5 @@ public class RemoteStoreIT extends RemoteStoreBaseIntegTestCase {
                 .setSettings(Settings.builder().put(IndexSettings.INDEX_TRANSLOG_DURABILITY_SETTING.getKey(), "async"))
                 .get()
         );
-    }
-
-    public void testCloseIndexWithNoOpSyncAndFlushForSyncTranslog() throws InterruptedException {
-        internalCluster().startNodes(3);
-        client().admin()
-            .cluster()
-            .prepareUpdateSettings()
-            .setTransientSettings(Settings.builder().put(CLUSTER_REMOTE_TRANSLOG_BUFFER_INTERVAL_SETTING.getKey(), "5s"))
-            .get();
-        Settings.Builder settings = Settings.builder()
-            .put(remoteStoreIndexSettings(0, 10000L, -1))
-            .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), "1s");
-        createIndex(INDEX_NAME, settings.build());
-        CountDownLatch latch = new CountDownLatch(1);
-        new Thread(() -> {
-            if (randomBoolean()) {
-                for (int i = 0; i < randomIntBetween(1, 5); i++) {
-                    indexSingleDoc(INDEX_NAME);
-                }
-                flushAndRefresh(INDEX_NAME);
-            }
-            // Index single doc to start the asyn io processor to run which will lead to 10s wait time before the next sync.
-            indexSingleDoc(INDEX_NAME);
-            // Reduce the latch for the main thread to flush after some sleep.
-            latch.countDown();
-            // Index another doc and in this case the flush would have happened before the sync.
-            indexSingleDoc(INDEX_NAME);
-        }).start();
-        // Wait for atleast one doc to be ingested.
-        latch.await();
-        // Sleep for some time for the next doc to be present in lucene buffer. If flush happens first before the doc #2
-        // gets indexed, then it goes into the happy case where the close index happens succefully.
-        Thread.sleep(1000);
-        // Flush so that the subsequent sync or flushes are no-op.
-        flush(INDEX_NAME);
-        // Closing the index involves translog.sync and shard.flush which are now no-op.
-        client().admin().indices().close(Requests.closeIndexRequest(INDEX_NAME)).actionGet();
-        Thread.sleep(10000);
-        ensureGreen(INDEX_NAME);
-    }
-
-    public void testCloseIndexWithNoOpSyncAndFlushForAsyncTranslog() throws InterruptedException {
-        internalCluster().startNodes(3);
-        Settings.Builder settings = Settings.builder()
-            .put(remoteStoreIndexSettings(0, 10000L, -1))
-            .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), "1s")
-            .put(IndexSettings.INDEX_TRANSLOG_DURABILITY_SETTING.getKey(), Durability.ASYNC)
-            .put(IndexSettings.INDEX_TRANSLOG_SYNC_INTERVAL_SETTING.getKey(), "10s");
-        createIndex(INDEX_NAME, settings.build());
-        CountDownLatch latch = new CountDownLatch(1);
-        new Thread(() -> {
-            // Index some docs to start the asyn io processor to run which will lead to 10s wait time before the next sync.
-            indexSingleDoc(INDEX_NAME);
-            indexSingleDoc(INDEX_NAME);
-            indexSingleDoc(INDEX_NAME);
-            // Reduce the latch for the main thread to flush after some sleep.
-            latch.countDown();
-        }).start();
-        // Wait for atleast one doc to be ingested.
-        latch.await();
-        // Flush so that the subsequent sync or flushes are no-op.
-        flush(INDEX_NAME);
-        // Closing the index involves translog.sync and shard.flush which are now no-op.
-        client().admin().indices().close(Requests.closeIndexRequest(INDEX_NAME)).actionGet();
-        Thread.sleep(10000);
-        ensureGreen(INDEX_NAME);
-    }
-
-    public void testLastSuccessfulFetchOfPinnedTimestampsPresentInNodeStats() throws Exception {
-        logger.info("Starting up cluster manager");
-        logger.info("cluster.remote_store.pinned_timestamps.enabled set to true");
-        logger.info("cluster.remote_store.pinned_timestamps.scheduler_interval set to minimum value of 1minute");
-        Settings pinnedTimestampEnabledSettings = Settings.builder()
-            .put(RemoteStoreSettings.CLUSTER_REMOTE_STORE_PINNED_TIMESTAMP_ENABLED.getKey(), true)
-            .put(RemoteStoreSettings.CLUSTER_REMOTE_STORE_PINNED_TIMESTAMP_SCHEDULER_INTERVAL.getKey(), "1m")
-            .build();
-        internalCluster().startClusterManagerOnlyNode(pinnedTimestampEnabledSettings);
-        internalCluster().startDataOnlyNode(pinnedTimestampEnabledSettings);
-        ensureStableCluster(2);
-
-        logger.info("Sleeping for 70 seconds to wait for fetching of pinned timestamps");
-        Thread.sleep(70000);
-
-        long lastSuccessfulFetchOfPinnedTimestamps = RemoteStorePinnedTimestampService.getPinnedTimestamps().v1();
-        assertTrue(lastSuccessfulFetchOfPinnedTimestamps > 0L);
-        assertBusy(() -> {
-            NodesStatsResponse nodesStatsResponse = internalCluster().client()
-                .admin()
-                .cluster()
-                .prepareNodesStats()
-                .addMetric(REMOTE_STORE_NODE_STATS.metricName())
-                .execute()
-                .actionGet();
-            for (NodeStats nodeStats : nodesStatsResponse.getNodes()) {
-                long lastRecordedFetch = nodeStats.getRemoteStoreNodeStats().getLastSuccessfulFetchOfPinnedTimestamps();
-                assertTrue(lastRecordedFetch >= lastSuccessfulFetchOfPinnedTimestamps);
-            }
-        }, 1, TimeUnit.MINUTES);
     }
 }
